requirements.txt
Django>=4.0
psycopg2-binary==2.9.3
djangorestframework==3.13.1
gunicorn==20.1.0
python-dotenv==0.19.2
django-filter==21.1
dj-database-url>=1.0.0
whitenoise>=6.0

